Crate telamon_kernels

Defines common kernels used to test and benchmark Telamon.

Modules

linalg
Linera algebra kernels.

statistics
Statistical analysis for benchmarking.

Traits

Kernel
A kernel that can be compiled, benchmarked and used for correctness tests.

Scalar
A scalar that can be used as the data type for tests.

Functions

analyze_bounds
Prints an analysis of the bounds computed by the lower bound model.