Module telamon_kernels::compose

Enums

ActivationFunction

Functions

matrix_matrix_multiply
Multiplies two matrices lhs and rhs

matrix_vector_multiply
Multiplies a matrix lhs with a vector rhs

tensor_add
Adds two tensors lhs and rhs of the same shape

tensor_elementwise_max
Applies the max function to all elements of a virtual tensor lhs with rhs as the second argument to max

tensor_elementwise_mul
Multiplies each element of a virtual tensor rhs with a scalar operand lhs

tensor_mad
Multiplies all elements of lhs_mul with rhs_mul_operand and adds the result to the tensor rhs_add

tensor_map
Opens dimensions mapped to the entire set of dimensions of a virtual tensor a and calls a function f with an operand representing a tensor's element and the builder. All further instructions created by f using the builder will be placed in a set of dimensions mapped to the dimensions of the virtual input tensor a.