Trait telamon_kernels::Kernel

pub trait Kernel: Sized {
    type Parameters;
    fn name() -> &'static str;

    fn build_signature(
        parameters: Self::Parameters, 
        is_generic: bool, 
        builder: &mut SignatureBuilder
    ) -> Self;

    fn build_body<'a>(
        &self, 
        signature: &'a Signature, 
        device: &'a Device
    ) -> Vec<SearchSpace<'a>>;

    fn check_result(&self, context: &Context) -> Result<(), String>;

    fn test(
        test_kind: TestKind, 
        params: Self::Parameters, 
        num_descend: usize, 
        context: &mut Context
    ) { ... }
}

A kernel that can be compiled, benchmarked and used for correctness tests.

Associated Types

type Parameters

The input parameters of the kernel.

Required Methods

fn name() -> &'static str

The name of the function computed by the kernel.

fn build_signature(
parameters: Self::Parameters,
is_generic: bool,
builder: &mut SignatureBuilder
) -> Self

Builds the signature of the kernel in the builder and returns an object that stores enough information to later build the kernel body and check its result. The is_generic flag indicates if th sizes should be instantiated.

fn build_body<'a>(
&self,
signature: &'a Signature,
device: &'a Device
) -> Vec<SearchSpace<'a>>

Builder the kernel body in the given builder. This builder should be based on the signature created by build_signature.

fn check_result(&self, context: &Context) -> Result<(), String>

Ensures the generated code performs the correct operation.

Provided Methods

fn test(
test_kind: TestKind,
params: Self::Parameters,
num_descend: usize,
context: &mut Context
)

Generates, executes and tests the output of candidates for the kernel.