Trait telamon_kernels::Kernel

pub trait Kernel<'a>: Sized {
    type Parameters: Clone + DeserializeOwned + Serialize;
    type ExpectedOutput;
    fn name() -> &'static str;

    fn build_signature<AM>(
        parameters: Self::Parameters, 
        builder: &mut SignatureBuilder<AM>
    ) -> Self
    where
        AM: ArgMap<'a> + Context;

    fn build_body<'b>(
        &self, 
        signature: &'b Signature, 
        ctx: &'b dyn Context
    ) -> Vec<Candidate<'b>>;

    fn get_expected_output(&self, _: &dyn Context) -> Self::ExpectedOutput;

    fn check_result(
        &self, 
        expected: &Self::ExpectedOutput, 
        context: &dyn Context
    ) -> Result<(), String>;

    fn generate_dump<'b, AM, F: Write>(
        params: Self::Parameters, 
        ctx: &'b mut AM, 
        sink: &mut F
    )
    where
        AM: Context + ArgMap<'a>,
    { ... }

    fn execute_dump<AM, F: Read>(ctx: &mut AM, dump: &mut F)
    where
        AM: Context + ArgMap<'a>,
    { ... }

    fn test_correctness<AM>(
        params: Self::Parameters, 
        num_tests: usize, 
        context: &mut AM
    )
    where
        AM: ArgMap<'a> + Context,
    { ... }

    fn test_bound<AM>(
        params: Self::Parameters, 
        num_tests: usize, 
        random_fill: bool, 
        context: &mut AM
    ) -> Vec<BoundSample>
    where
        AM: ArgMap<'a> + Context,
    { ... }

    fn benchmark<AM>(
        config: &Config, 
        params: Self::Parameters, 
        num_samples: usize, 
        random_fill: bool, 
        context: &mut AM
    ) -> Vec<f64>
    where
        AM: ArgMap<'a> + Context,
    { ... }

    fn deadend_ratio<AM>(
        params: Self::Parameters, 
        num_samples: usize, 
        context: &mut AM
    ) -> f64
    where
        AM: ArgMap<'a> + Context,
    { ... }
}

A kernel that can be compiled, benchmarked and used for correctness tests.

Associated Types

type Parameters: Clone + DeserializeOwned + Serialize

The input parameters of the kernel.

type ExpectedOutput

The values to expect as output.

Loading content...
Required methods

fn name() -> &'static str

The name of the function computed by the kernel.

fn build_signature<AM>(
parameters: Self::Parameters,
builder: &mut SignatureBuilder<AM>
) -> Self where
AM: ArgMap<'a> + Context,

Builds the signature of the kernel in the builder and returns an object that stores enough information to later build the kernel body and check its result.

fn build_body<'b>(
&self,
signature: &'b Signature,
ctx: &'b dyn Context
) -> Vec<Candidate<'b>>

Builder the kernel body in the given builder. This builder should be based on the signature created by build_signature.

fn get_expected_output(&self, _: &dyn Context) -> Self::ExpectedOutput

Computes the expected output.

fn check_result(
&self,
expected: &Self::ExpectedOutput,
context: &dyn Context
) -> Result<(), String>

Ensures the generated code performs the correct operation.

Loading content...
Provided methods

fn generate_dump<'b, AM, F: Write>(
params: Self::Parameters,
ctx: &'b mut AM,
sink: &mut F
) where
AM: Context + ArgMap<'a>,

Generate a dump of a specific implementation of Self in a file, so we can rerun tests on the same candidate multiple times. More specifically, we dump the list of actions taken on the candidate rather than the candidate itself

fn execute_dump<AM, F: Read>(ctx: &mut AM, dump: &mut F) where
AM: Context + ArgMap<'a>,

Takes a path to a log and execute it. Caller is responsible for making sure that the log corresponds to the kernel and context being executed

fn test_correctness<AM>(
params: Self::Parameters,
num_tests: usize,
context: &mut AM
) where
AM: ArgMap<'a> + Context,

Generates, executes and tests the output of candidates for the kernel.

fn test_bound<AM>(
params: Self::Parameters,
num_tests: usize,
random_fill: bool,
context: &mut AM
) -> Vec<BoundSample> where
AM: ArgMap<'a> + Context,

Tests the correctness of the bound of kernels and returns the list of tested leafs along with the actual evaluation time.

fn benchmark<AM>(
config: &Config,
params: Self::Parameters,
num_samples: usize,
random_fill: bool,
context: &mut AM
) -> Vec<f64> where
AM: ArgMap<'a> + Context,

Runs the search and benchmarks the resulting candidate.

fn deadend_ratio<AM>(
params: Self::Parameters,
num_samples: usize,
context: &mut AM
) -> f64 where
AM: ArgMap<'a> + Context,

Computes the probability of encountering a dead-end when descending in the search tree.

Loading content...